+++ src/test/java/com/selina/lending/internal/mapper/MapperBase.java
@@ -121,7 +121,7 @@ public abstract class MapperBase {
     public static final Double INCOME_AMOUNT = 15000.00;
     public static final String INCOME_TYPE = "Gross salary";
     public static final String LOAN_PURPOSE = "Home improvements";
-    public static final Integer LOAN_AMOUNT = 50000;
+    public static final Double LOAN_AMOUNT = 50000.00;
     public static final int LOAN_TERM = 5;
     public static final double ALLOCATION_AMOUNT = 50000;
     public static final Double AFFORDABILITY_DEFICIT = 2100.0;
